Update: Move DataAccess/DependencyInjection to the right place
